+++ docs/superpowers/specs/2026-08-10-plugin-wechat-channels-decrypt-utils-design.md
@@ -1,0 +1,43 @@
+# 视频号媒体解密纯工具模块拆分设计
+
+## 背景
+
+插件 `src/main.js` 同时承载业务编排和一段自包含的 ISAAC-64 解密算法。该算法只接收解密键、Buffer 和长度上限，不读取设置、网络、文件或账户状态，适合从主文件中独立出来，降低主文件体积和后续定位成本。
+
+## 边界
+
+迁移范围严格限定为：
+
+- 默认加密头长度常量 `131072`；
+- `u64`、`Isaac64` 和解密键解析内部实现；
+- `generateWechatChannelsDecryptorBytes`；
+- `decryptWechatChannelsMediaBuffer`。
+
+主文件继续负责媒体下载、候选地址、解密键来源、有效媒体校验和后续转写。调用点与测试辅助导出保持原名。
+
+## 方案比较
+
+### 方案 A：原样迁移为 CommonJS 纯工具模块（采用）
+
+把完整算法逐字迁移到 `src/wechat-channels-decrypt-utils.js`，只导出两个现有公共函数。主文件以相对路径导入。优点是依赖方向单一、可独立固定向量测试、行为面最小。
+
+### 方案 B：同时重写或优化 ISAAC-64
+
+可能减少代码，但会改变位运算、字节序或 BigInt 行为，无法证明与已发布媒体兼容，拒绝。
+
+### 方案 C：只包一层代理，保留算法在主文件
+
+不能减少主文件耦合和体积，也无法形成独立测试边界，拒绝。
+
+## 等价性证明
+
+1. 固定 seed `123456789` 的前 32 字节必须为 `9c87e6946a410182aee08a5a116f7cb986f10740dd490b9a330cd31db53ddc4a`。
+2. 十进制 `123456789` 与十六进制 `0x75BCD15` 生成完全相同的密钥流。
+3. 无效键生成空密钥流；解密时原样复制输入。
+4. 默认只异或前 131072 字节，后续字节不变；自定义 limit 只影响指定前缀。
+5. 同一键对结果再次解密必须还原原始 Buffer。
+6. 生成后的单文件 bundle 需通过现有视频号下载路径测试和双构建哈希比较。
+
+## 风险与回退
+
+风险集中在 ISAAC-64 状态初始化、64 位截断、低字节输出和解密长度。任何固定向量、基线回归或双构建哈希不一致即停止，丢弃本阶段 Worktree；第七阶段提交保持可直接回退。
